Little generator cleanup
